requirements.txt
adrf==0.1.1
aiogram==2.25.1
aioimaplib==1.0.1
anyio==3.6.2
celery~=5.2.7
Django==4.1.7
django-celery-beat==2.5.0
djangorestframework==3.14.0
httpx==0.23.3
pbkdf2==1.3
playwright==1.31.1
pre-commit==3.1.1
psycopg2-binary==2.9.5
pytest==7.2.2
pytest-asyncio==0.21.0
python-dotenv==1.0.0
rsa==4.9
redis==4.5.1
sentry-sdk[django]
uvicorn==0.20.0
virtualenv==20.20.0
whitenoise==6.4.0
drf_spectacular